backend/pyproject.toml
[tool.poetry]
name = "replinet-backend"
version = "0.1.0"
description = "Backend for Replinet - AI Agents Management Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
redis = {extras = ["hiredis"], version = "^5.0.1"}
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
httpx = "^0.25.2"
beautifulsoup4 = "^4.12.2"
playwright = "^1.40.0"
websockets = "^12.0.0"
python-dateutil = "^2.8.2"
pillow = "^10.1.0"
celery = "^5.3.6"
gunicorn = "^21.2.0"
docker = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
ruff = "^0.1.6"
pre-commit = "^3.5.0"
httpx = "^0.25.2"
pytest-env = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
disallow_incomplete_defs = true
disallow_untyped_calls = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=app --cov-report=term-missing"
env = [
    "DATABASE_URL=postgresql://postgres:replinet@localhost:5432/replinet_test",
    "REDIS_URL=redis://localhost:6379/1",
    "SECRET_KEY=test_secret_key",
    "ENVIRONMENT=test",
]

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]